@@ -1,8 +1,8 @@
 package com.github.tvbox.osc.ui.adapter;
 
-import android.graphics.Color;
 import android.widget.TextView;
 
+import com.blankj.utilcode.util.ColorUtils;
 import com.chad.library.adapter.base.BaseQuickAdapter;
 import com.chad.library.adapter.base.BaseViewHolder;
 import com.github.tvbox.osc.R;
@@ -32,7 +32,7 @@ public class LiveSettingGroupAdapter extends BaseQuickAdapter<LiveSettingGroup, 
         if (groupIndex == selectedGroupIndex && groupIndex != focusedGroupIndex) {
             tvGroupName.setTextColor(mContext.getResources().getColor(R.color.color_1890FF));
         } else {
-            tvGroupName.setTextColor(Color.WHITE);
+            tvGroupName.setTextColor(ColorUtils.getColor(R.color.text_foreground));
         }
     }
 
