@@ -20,15 +20,11 @@ import me.jessyan.autosize.utils.AutoSizeUtils;
 
 public class SearchAdapter extends BaseQuickAdapter<Movie.Video, BaseViewHolder> {
     public SearchAdapter() {
-        super(Hawk.get(HawkConfig.SEARCH_VIEW, 0) == 0 ? R.layout.item_search_lite : R.layout.item_search, new ArrayList<>());
+        super(R.layout.item_search, new ArrayList<>());
     }
 
     @Override
     protected void convert(BaseViewHolder helper, Movie.Video item) {
-        // lite
-        if (Hawk.get(HawkConfig.SEARCH_VIEW, 0) == 0) {
-            helper.setText(R.id.tvName, String.format("%s  %s %s %s", ApiConfig.get().getSource(item.sourceKey).getName(), item.name, item.type == null ? "" : item.type, item.note == null ? "" : item.note));
-        } else {// with preview
             helper.setText(R.id.tvName, item.name);
             helper.setText(R.id.tvSite, ApiConfig.get().getSource(item.sourceKey).getName());
             helper.setVisible(R.id.tvNote, item.note != null && !item.note.isEmpty());
@@ -41,14 +37,13 @@ public class SearchAdapter extends BaseQuickAdapter<Movie.Video, BaseViewHolder>
                         .load(item.pic)
                         .transform(new RoundTransformation(MD5.string2MD5(item.pic + "position=" + helper.getLayoutPosition()))
                                 .centerCorp(true)
-                                .override(AutoSizeUtils.mm2px(mContext, 300), AutoSizeUtils.mm2px(mContext, 400))
-                                .roundRadius(AutoSizeUtils.mm2px(mContext, 10), RoundTransformation.RoundType.ALL))
+                                .override(AutoSizeUtils.dp2px(mContext, 100), AutoSizeUtils.dp2px(mContext, 140))
+                                .roundRadius(AutoSizeUtils.dp2px(mContext, 20), RoundTransformation.RoundType.ALL))
                         .placeholder(R.drawable.img_loading_placeholder)
                         .error(R.drawable.img_loading_placeholder)
                         .into(ivThumb);
             } else {
                 ivThumb.setImageResource(R.drawable.img_loading_placeholder);
             }
-        }
     }
 }
