@@ -1,12 +1,18 @@
 package com.github.tvbox.osc.ui.adapter;
 
 import android.graphics.Color;
+import android.view.ViewGroup;
 import android.widget.TextView;
 
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.blankj.utilcode.util.ConvertUtils;
 import com.chad.library.adapter.base.BaseQuickAdapter;
 import com.chad.library.adapter.base.BaseViewHolder;
 import com.github.tvbox.osc.R;
 import com.github.tvbox.osc.bean.VodInfo;
+import com.lihang.ShadowLayout;
 
 import java.util.ArrayList;
 
@@ -16,18 +22,29 @@ import java.util.ArrayList;
  * @description:
  */
 public class SeriesAdapter extends BaseQuickAdapter<VodInfo.VodSeries, BaseViewHolder> {
-    public SeriesAdapter() {
+    private boolean isGird;
+
+    public SeriesAdapter(boolean isGird) {
         super(R.layout.item_series, new ArrayList<>());
+        this.isGird = isGird;
     }
 
     @Override
     protected void convert(BaseViewHolder helper, VodInfo.VodSeries item) {
+        ShadowLayout sl = helper.getView(R.id.sl);
         TextView tvSeries = helper.getView(R.id.tvSeries);
-        if (item.selected) {
-            tvSeries.setTextColor(mContext.getResources().getColor(R.color.color_02F8E1));
-        } else {
-            tvSeries.setTextColor(Color.WHITE);
+        sl.setSelected(item.selected);
+        tvSeries.setText(item.name);
+
+        if (!isGird){// 详情页横向展示时固定宽度
+            ViewGroup.LayoutParams layoutParams = sl.getLayoutParams();
+            layoutParams.width = ConvertUtils.dp2px(120);
+            sl.setLayoutParams(layoutParams);
         }
-        helper.setText(R.id.tvSeries, item.name);
     }
+
+    public void setGird(boolean gird) {
+        isGird = gird;
+    }
+
 }
