@@ -17,18 +17,17 @@ import java.util.ArrayList;
  */
 public class SeriesFlagAdapter extends BaseQuickAdapter<VodInfo.VodSeriesFlag, BaseViewHolder> {
     public SeriesFlagAdapter() {
-        super(R.layout.item_series_flag, new ArrayList<>());
+        super(R.layout.item_select_flag, new ArrayList<>());
     }
 
     @Override
     protected void convert(BaseViewHolder helper, VodInfo.VodSeriesFlag item) {
-        TextView tvSeries = helper.getView(R.id.tvSeriesFlag);
-        View select = helper.getView(R.id.tvSeriesFlagSelect);
+        View select = helper.getView(R.id.vFlag);
         if (item.selected) {
             select.setVisibility(View.VISIBLE);
         } else {
             select.setVisibility(View.GONE);
         }
-        helper.setText(R.id.tvSeriesFlag, item.name);
+        helper.setText(R.id.tvFlag, item.name);
     }
 }
